refactor Remove redundant calls of the conversion method.

+++ src/main/kotlin/jp/toastkid/yobidashi/compact/calendar/service/EquinoxDayCalculator.kt
@@ -3,11 +3,11 @@ package jp.toastkid.yobidashi.compact.calendar.service
 class EquinoxDayCalculator {
 
     fun calculateVernalEquinoxDay(year: Int): Int {
-        return (20.8431 + 0.242194 * (year - 1980)).toInt() - ((year - 1980) / 4).toInt()
+        return (20.8431 + 0.242194 * (year - 1980)).toInt() - ((year - 1980) / 4)
     }
 
     fun calculateAutumnalEquinoxDay(year: Int): Int {
-        return (23.2488 + 0.242194 * (year - 1980)).toInt() - ((year - 1980) / 4).toInt()
+        return (23.2488 + 0.242194 * (year - 1980)).toInt() - ((year - 1980) / 4)
     }
 
 }
